+++ packages/plugin-hardhat/src/index.ts
@@ -22,6 +22,11 @@ export interface HardhatUpgrades {
     changeProxyAdmin: ChangeAdminFunction;
     transferProxyAdminOwnership: TransferProxyAdminOwnershipFunction;
   };
+  erc1967: {
+    getAdminAddress: (proxyAdress: string) => Promise<string>;
+    getImplementationAddress: (proxyAdress: string) => Promise<string>;
+    getBeaconAddress: (proxyAdress: string) => Promise<string>;
+  };
 }
 
 interface RunCompilerArgs {
@@ -63,7 +68,12 @@ subtask(TASK_COMPILE_SOLIDITY_COMPILE, async (args: RunCompilerArgs, hre, runSup
 
 extendEnvironment(hre => {
   hre.upgrades = lazyObject((): HardhatUpgrades => {
-    const { silenceWarnings } = require('@openzeppelin/upgrades-core');
+    const {
+      silenceWarnings,
+      getAdminAddress,
+      getImplementationAddress,
+      getBeaconAddress,
+    } = require('@openzeppelin/upgrades-core');
     const { makeDeployProxy } = require('./deploy-proxy');
     const { makeUpgradeProxy } = require('./upgrade-proxy');
     const { makePrepareUpgrade } = require('./prepare-upgrade');
@@ -78,6 +88,11 @@ extendEnvironment(hre => {
         getInstance: makeGetInstanceFunction(hre),
         changeProxyAdmin: makeChangeProxyAdmin(hre),
         transferProxyAdminOwnership: makeTransferProxyAdminOwnership(hre),
+      },
+      erc1967: {
+        getAdminAddress: proxyAddress => getAdminAddress(hre.network.provider, proxyAddress),
+        getImplementationAddress: proxyAddress => getImplementationAddress(hre.network.provider, proxyAddress),
+        getBeaconAddress: proxyAddress => getBeaconAddress(hre.network.provider, proxyAddress),
       },
     };
   });