+++ packages/plugin-truffle/src/index.ts
@@ -4,3 +4,4 @@ export * from './deploy-proxy';
 export * from './prepare-upgrade';
 export * from './upgrade-proxy';
 export { admin } from './admin';
+export { erc1967 } from './erc1967';